+++ src/app/(client)/meal/post/_components/add-meal-drawer.tsx
@@ -4,9 +4,9 @@ import { Button } from '@/components/ui/button';
 import { Drawer, DrawerClose, DrawerContent, DrawerFooter, DrawerTitle, DrawerTrigger } from '@/components/ui/drawer';
 import { FormControl, FormField, FormItem, FormLabel, Form, FormMessage } from '@/components/ui/form';
 import { Input } from '@/components/ui/input';
-import { MAX_MENU_NAME_LENGTH, MAX_NUMERIC_LENGTH } from '../edit/constants/meal-edit.constant';
+import { MAX_MENU_NAME_LENGTH, MAX_NUMERIC_LENGTH } from '../edit/_constants/meal-edit.constant';
 import { formatNumberWithComma } from '@/utils/format.util';
-import { parseNumber } from '../edit/utils/meal-edit.util';
+import { parseNumber } from '../edit/_utils/meal-edit.util';
 import { useForm } from 'react-hook-form';
 import { useState } from 'react';
 import { generateCaloriesAnalysisByText } from '@/apis/gemini.api';
