+++ src/app/(client)/meal/post/_components/add-meal-modal.tsx
@@ -7,9 +7,9 @@ import { Input } from '@/components/ui/input';
 import { Typography } from '@/components/ui/typography';
 import { Dialog, DialogClose, DialogContent, DialogOverlay, DialogTitle, DialogTrigger } from '@radix-ui/react-dialog';
 import { useForm } from 'react-hook-form';
-import { MAX_MENU_NAME_LENGTH, MAX_NUMERIC_LENGTH } from '../edit/constants/meal-edit.constant';
+import { MAX_MENU_NAME_LENGTH, MAX_NUMERIC_LENGTH } from '../edit/_constants/meal-edit.constant';
 import { formatNumberWithComma } from '@/utils/format.util';
-import { parseNumber } from '../edit/utils/meal-edit.util';
+import { parseNumber } from '../edit/_utils/meal-edit.util';
 import { useRouter } from 'next/navigation';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { createAiRequestByText, createAiResponse } from '@/apis/analysis-request.api';
